GRAILS-5329 - Add the test phase and type names to report files to avoid naming conflicts.
This effectively solves the issue, but it's a little suboptimal getting this stuff from the binding. However, this is the least
intrusive way to solve this issue right now and I am reluctant to make any modifications to GrailsTestType to support this. In particular,
it would be necessary for test types to have knowledge of the phase they are in. There is no obvious way to add this to the API/contract.